cast in Fixed_cmp_dist

+++ Packages/Fixed_precision_nt/src/Fixed_precision_nt.C
@@ -260,10 +260,10 @@ int Fixed_cmp_dist(float x0, float y0,
 		   float x1, float y1, 
 		   float x2, float y2)
 {
-  double X1 = x1-x0;
-  double Y1 = y1-y0;
-  double X2 = x2-x0;
-  double Y2 = y2-y0;
+  double X1 = (double)x1-(double)x0;
+  double Y1 = (double)y1-(double)y0;
+  double X2 = (double)x2-(double)x0;
+  double Y2 = (double)y2-(double)y0;
   X1 = (X2*X2+Y2*Y2)-(X1*X1+Y1*Y1);         // exact
   return (X1>0) ? 1 : (X1==0) ? 0 : -1;
 }
@@ -273,12 +273,12 @@ int Fixed_cmp_dist(float x0, float y0, float z0,
 		   float x1, float y1, float z1,
 		   float x2, float y2, float z2)
 {
-  double X1 = x1-x0;
-  double Y1 = y1-y0;
-  double Z1 = z1-z0;
-  double X2 = x2-x0;
-  double Y2 = y2-y0;
-  double Z2 = z2-z0;
+  double X1 = (double)x1-(double)x0;
+  double Y1 = (double)y1-(double)y0;
+  double Z1 = (double)z1-(double)z0;
+  double X2 = (double)x2-(double)x0;
+  double Y2 = (double)y2-(double)y0;
+  double Z2 = (double)z2-(double)z0;
   X1 = (X2*X2+Y2*Y2+Z2*Z2)-(X1*X1+Y1*Y1+Z1*Z1);         // exact
   return (X1>0) ? 1 : (X1==0) ? 0 : -1;
 }
